+++ src/applications/caregivers/routes.jsx
@@ -1,4 +1,4 @@
-import { createRoutes } from 'platform/forms-system/src/js/routing';
+import { createRoutes } from 'platform/forms-system/src/js/routing/createRoutes';
 import formConfig from './config/form';
 import App from './containers/App';
