+++ src/platform/forms-system/src/js/containers/FormPage.jsx
@@ -23,6 +23,7 @@ import {
   getNextPagePath,
   getPreviousPagePath,
   checkValidPagePath,
+  goBack,
 } from '../routing';
 import { DevModeNavLinks } from '../components/dev/DevModeNavLinks';
 import { stringifyUrlParams } from '../helpers';
@@ -206,32 +207,15 @@ class FormPage extends React.Component {
   };
 
   goBack = () => {
-    const { form, route, location } = this.props;
-
-    const path = getPreviousPagePath(
-      route.pageList,
-      form.data,
-      location.pathname,
-    );
-
-    if (typeof route.pageConfig.onNavBack === 'function') {
-      route.pageConfig.onNavBack({
-        formData: form.data,
-        goPath: customPath => this.props.router.push(customPath),
-        goPreviousPath: urlParams => {
-          const urlParamsString = stringifyUrlParams(urlParams);
-          this.props.router.push(path + (urlParamsString || ''));
-        },
-        pageList: route.pageList,
-        pathname: location.pathname,
-        setFormData: this.props.setData,
-        urlParams: location.query,
-        index: this.props.params?.index,
-      });
-      return;
-    }
-
-    this.props.router.push(path);
+    goBack({
+      formData: this.formData(),
+      index: this.props.params?.index,
+      location: this.props.location,
+      onNavBack: this.props.route.pageConfig?.onNavBack,
+      router: this.props.router,
+      setData: this.props.setData,
+      pageList: this.props.route?.pageList,
+    });
   };
 
   goToPath = (customPath, options = {}) => {
