+++ src/platform/forms/save-in-progress/helpers.js
@@ -2,7 +2,7 @@ import {
   createPageList,
   createFormPageList,
 } from 'platform/forms-system/src/js/helpers';
-import { createRoutes } from 'platform/forms-system/src/js/routing';
+import { createRoutes } from 'platform/forms-system/src/js/routing/createRoutes';
 import RoutedSavablePage from './RoutedSavablePage';
 import RoutedSavableReviewPage from './RoutedSavableReviewPage';
 import FormSaved from './FormSaved';
